brisk/requirements.txt
streamlit==1.46.1
google-generativeai==0.8.5
python-dotenv==1.1.1 